pyproject.toml
[tool.poetry]
name = "patchcore"
version = "0.1.0"
description = ""
authors = ["gengen"]
readme = "README.md"
packages = [{include = "patchcore"}]

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.3.0"
torchvision = "^0.18.0"
numpy = "^1.26.4"
scikit-learn = "^1.5.0"
opencv-python = "^4.9.0.80"
scipy = "^1.13.1"
pillow = "^10.3.0"
tqdm = "^4.66.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py310']
include = '\.py$'
skip-string-normalization = true
exclude = '''
(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | venv
    | \.idea
    | \.vscode
    | _build
    | buck-out
    | build
    | dist
    | migrations
)
'''

[tool.isort]
profile = "black"
line_length = 120
skip_glob = "*/migrations/*.py"
